# backend/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sample Todo API",
    "description": "Sample todo backend for a senior capstone class. Demonstrates CRUD, auth (Firebase), error handling, and database integration.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/signin": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Authenticate with a Firebase ID token",
        "description": "Exchanges a Firebase ID token (minted by the client's Firebase SDK) for a local user record. The user is created on first sign-in. Returns the same token to be sent as `Authorization: Bearer <token>` on all protected routes.",
        "operationId": "signin_api_v1_auth_signin_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignInResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/users/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get the current authenticated user",
        "description": "Returns the profile of the user identified by the Bearer token.",
        "operationId": "read_current_user_api_v1_auth_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/todos": {
      "get": {
        "tags": [
          "todos"
        ],
        "summary": "List all todos",
        "description": "Returns every todo owned by the authenticated user.",
        "operationId": "read_all_todos_api_v1_todos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TodoOut"
                  },
                  "type": "array",
                  "title": "Response Read All Todos Api V1 Todos Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "todos"
        ],
        "summary": "Create a todo",
        "description": "Creates a new todo belonging to the authenticated user.",
        "operationId": "create_new_todo_api_v1_todos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/todos/{todo_id}": {
      "get": {
        "tags": [
          "todos"
        ],
        "summary": "Get a todo by id",
        "description": "Returns a single todo if it exists and belongs to the user, else 404.",
        "operationId": "read_todo_api_v1_todos__todo_id__get",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "todo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Todo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "todos"
        ],
        "summary": "Update a todo by id",
        "description": "Updates the fields of a todo. Only non-null fields are changed; passing null leaves a field unchanged.",
        "operationId": "update_existing_todo_api_v1_todos__todo_id__put",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "todo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Todo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "todos"
        ],
        "summary": "Delete a todo by id",
        "description": "Deletes a todo. Returns 204 on success, 404 if not found.",
        "operationId": "delete_existing_todo_api_v1_todos__todo_id__delete",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "todo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Todo Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Health check",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SignInRequest": {
        "properties": {
          "id_token": {
            "type": "string",
            "title": "Id Token",
            "description": "Firebase ID token from the client's auth SDK"
          }
        },
        "type": "object",
        "required": [
          "id_token"
        ],
        "title": "SignInRequest",
        "description": "Body for /auth/signin: a Firebase ID token minted by the client."
      },
      "SignInResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "Firebase ID token echoed back for the client"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user"
        ],
        "title": "SignInResponse"
      },
      "TodoCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "TodoCreate"
      },
      "TodoOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "completed": {
            "type": "boolean",
            "title": "Completed"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "completed",
          "owner_id",
          "created_at",
          "updated_at"
        ],
        "title": "TodoOut"
      },
      "TodoUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "completed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed"
          }
        },
        "type": "object",
        "title": "TodoUpdate"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "UserOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase ID token. Obtain it from `POST /auth/signin`."
      }
    }
  },
  "tags": [
    {
      "name": "auth",
      "description": "Authentication with Firebase ID tokens"
    },
    {
      "name": "todos",
      "description": "Todo CRUD operations"
    }
  ]
}