# backend/tests-integration/todo_api_client/models/sign_in_response.py
# coding: utf-8

"""
    Sample Todo API

    Sample todo backend for a senior capstone class. Demonstrates CRUD, auth (Firebase), error handling, and database integration.

    The version of the OpenAPI document: 0.1.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json

from pydantic import BaseModel, ConfigDict, Field, StrictStr
from typing import Any, ClassVar, Dict, List, Optional
from todo_api_client.models.user_out import UserOut
from typing import Optional, Set
from typing_extensions import Self
from pydantic_core import to_jsonable_python

class SignInResponse(BaseModel):
    """
    SignInResponse
    """ # noqa: E501
    access_token: StrictStr = Field(description="Firebase ID token echoed back for the client")
    token_type: Optional[StrictStr] = 'bearer'
    user: UserOut
    __properties: ClassVar[List[str]] = ["access_token", "token_type", "user"]

    model_config = ConfigDict(
        validate_by_name=True,
        validate_by_alias=True,
        validate_assignment=True,
        protected_namespaces=(),
    )


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        return json.dumps(to_jsonable_python(self.to_dict()))

    @classmethod
    def from_json(cls, json_str: str) -> Optional[Self]:
        """Create an instance of SignInResponse from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        excluded_fields: Set[str] = set([
        ])

        _dict = self.model_dump(
            by_alias=True,
            exclude=excluded_fields,
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of user
        if self.user:
            _dict['user'] = self.user.to_dict()
        return _dict

    @classmethod
    def from_dict(cls, obj: Optional[Dict[str, Any]]) -> Optional[Self]:
        """Create an instance of SignInResponse from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "access_token": obj.get("access_token"),
            "token_type": obj.get("token_type") if obj.get("token_type") is not None else 'bearer',
            "user": UserOut.from_dict(obj["user"]) if obj.get("user") is not None else None
        })
        return _obj